app-pi4-transmitter/src/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-rfm9x==2.2.17
adafruit-circuitpython-typing==1.11.2
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
azure-cognitiveservices-speech==1.41.1
binho-host-adapter==0.1.6
board==1.0
colorzero==2.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
gpiozero==2.0.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.10.0
noisereduce==3.0.3
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
psutil==6.1.0
PyAudio==0.2.14
pyftdi==0.56.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
scipy==1.14.1
six==1.17.0
sysv-ipc==1.1.0
tqdm==4.67.1
typing_extensions==4.12.2
